Stop catching ActiveRecord::RecordNotUnique
To try to save a read before writing to the DB, I had been proactively
trying to run a `create` on the various cruft_tracker records. I'd then
catch any ActiveRecord::RecordNotUnique and create a record, since it
didn't already exist. This resulted in spamming RecordNotUnique errors
and really wasn't helpful anyhow. This commit just tries reading the
record first with find_by. If not found, the record is created.